Move all kernel_tests in individual ops to tests
So that we could use `(cd tests && python -m pytest -s .)`
