+++ tensorflow_io/arrow/BUILD
@@ -18,39 +18,3 @@ cc_binary(
     ],
     copts = ["-pthread", "-std=c++11", "-D_GLIBCXX_USE_CXX11_ABI=0", "-DNDEBUG"]
 )
-
-py_library(
-    name = "arrow_ops_py",
-    srcs = [
-        "python/ops/arrow_dataset_ops.py",
-    ],
-    data = [
-        ":python/ops/_arrow_ops.so"
-    ],
-    srcs_version = "PY2AND3",
-)
-
-py_test(
-    name = "arrow_py_test",
-    srcs = [
-        "python/kernel_tests/arrow_test.py"
-    ],
-    main = "python/kernel_tests/arrow_test.py",
-    deps = [
-        ":arrow_ops_py"
-    ],
-    srcs_version = "PY2AND3",
-)
-
-py_library(
-    name = "arrow_py",
-    srcs = [
-        "__init__.py",
-        "python/__init__.py",
-        "python/ops/__init__.py",
-    ],
-    deps = [
-        ":arrow_ops_py",
-    ],
-    srcs_version = "PY2AND3",
-)
