+++ tensorflow_io/ignite/BUILD
@@ -41,71 +41,3 @@ cc_binary(
     ],
     copts = ["-pthread", "-std=c++11", "-D_GLIBCXX_USE_CXX11_ABI=0", "-DNDEBUG"]
 )
-
-py_library(
-    name = "dataset_ops_py",
-    srcs = [
-        "python/ops/ignite_dataset_ops.py",
-    ],
-    data = [
-        ":python/ops/_ignite_ops.so"
-    ],
-    srcs_version = "PY2AND3",
-)
-
-py_library(
-    name = "igfs_ops_py",
-    srcs = [
-        "python/ops/igfs_ops.py",
-    ],
-    data = [
-        ":python/ops/_ignite_ops.so"
-    ],
-    srcs_version = "PY2AND3",
-)
-
-# The Apache Ignite servers have to setup before the test and tear down
-# after the test manually. The docker engine has to be installed.
-#
-# To setup Apache Ignite servers:
-# $ bash ./python/tests/start_ignite.sh
-#
-# To tear down Apache Ignite servers:
-# $ bash ./python/tests/stop_ignite.sh
-py_test(
-    name = "ignite_py_test",
-    srcs = [
-        "python/tests/ignite_dataset_test.py",
-    ],
-    main = "python/tests/ignite_dataset_test.py",
-    deps = [
-        ":ignite_py",
-    ],
-    srcs_version = "PY2AND3",
-)
-
-py_test(
-    name = "igfs_py_test",
-    srcs = [
-        "python/tests/igfs_test.py",
-    ],
-    main = "python/tests/igfs_test.py",
-    deps = [
-        ":ignite_py",
-    ],
-    srcs_version = "PY2AND3",
-)
-
-py_library(
-    name = "ignite_py",
-    srcs = [
-        "__init__.py",
-        "python/__init__.py",
-        "python/ops/__init__.py",
-    ],
-    deps = [
-        ":dataset_ops_py",
-        ":igfs_ops_py",
-    ],
-    srcs_version = "PY2AND3",
-)
