+++ tensorflow_io/kafka/BUILD
@@ -16,39 +16,3 @@ cc_binary(
     ],
     copts = ["-pthread", "-std=c++11", "-D_GLIBCXX_USE_CXX11_ABI=0", "-DNDEBUG"]
 )
-
-py_library(
-    name = "kafka_ops_py",
-    srcs = ([
-        "python/ops/kafka_dataset_ops.py",
-    ]),
-    data = [
-        ":python/ops/_kafka_ops.so"
-    ],
-    srcs_version = "PY2AND3",
-)
-
-py_test(
-    name = "kafka_py_test",
-    srcs = [
-        "python/kernel_tests/kafka_test.py"
-    ],
-    main = "python/kernel_tests/kafka_test.py",
-    deps = [
-        ":kafka_ops_py",
-    ],
-    srcs_version = "PY2AND3",
-)
-
-py_library(
-    name = "kafka_py",
-    srcs = ([
-        "__init__.py",
-        "python/__init__.py",
-        "python/ops/__init__.py",
-    ]),
-    deps = [
-        ":kafka_ops_py"
-    ],
-    srcs_version = "PY2AND3",
-)
