+++ tensorflow_io/parquet/BUILD
@@ -2,11 +2,6 @@ licenses(["notice"])  # Apache 2.0
 
 package(default_visibility = ["//visibility:public"])
 
-filegroup(
-    name = "test_data",
-    srcs = glob(["python/kernel_tests/testdata/*"]),
-)
-
 cc_binary(
     name = 'python/ops/_parquet_ops.so',
     srcs = [
@@ -20,43 +15,4 @@ cc_binary(
         "@arrow//:arrow",
     ],
     copts = ["-pthread", "-std=c++11", "-D_GLIBCXX_USE_CXX11_ABI=0", "-DNDEBUG"]
-)
-
-py_library(
-    name = "parquet_ops_py",
-    srcs = ([
-        "python/ops/parquet_dataset_ops.py",
-    ]),
-    data = [
-        ":python/ops/_parquet_ops.so"
-    ],
-    srcs_version = "PY2AND3",
-)
-
-py_test(
-    name = "parquet_py_test",
-    srcs = [
-        "python/kernel_tests/parquet_test.py"
-    ],
-    main = "python/kernel_tests/parquet_test.py",
-    data = [
-        ":test_data",
-    ],
-    deps = [
-        ":parquet_ops_py",
-    ],
-    srcs_version = "PY2AND3",
-)
-
-py_library(
-    name = "parquet_py",
-    srcs = ([
-        "__init__.py",
-        "python/__init__.py",
-        "python/ops/__init__.py",
-    ]),
-    deps = [
-        ":parquet_ops_py"
-    ],
-    srcs_version = "PY2AND3",
 )
