+++ tensorflow_io/video/BUILD
@@ -2,11 +2,6 @@ licenses(["notice"])  # Apache 2.0
 
 package(default_visibility = ["//visibility:public"])
 
-filegroup(
-    name = "test_data",
-    srcs = glob(["python/kernel_tests/testdata/*"]),
-)
-
 # Note: stub files are generated from Ubuntu 14.0 with
 # `echo '' | g++ -shared -fPIC -x c++ - -o $@`
 filegroup(
@@ -96,45 +91,4 @@ cc_binary(
         ":stub.so",
     ],
     copts = ["-pthread", "-std=c++11", "-D_GLIBCXX_USE_CXX11_ABI=0", "-DNDEBUG"]
-)
-
-py_library(
-    name = "video_ops_py",
-    srcs = ([
-        "python/ops/video_dataset_ops.py",
-    ]),
-    data = [
-        ":python/ops/_video_ops_ffmpeg_3.4.so",
-        ":python/ops/_video_ops_ffmpeg_2.8.so",
-        ":python/ops/_video_ops_libav_9.20.so",
-    ],
-    srcs_version = "PY2AND3",
-)
-
-py_test(
-    name = "video_py_test",
-    srcs = [
-        "python/kernel_tests/video_test.py"
-    ],
-    main = "python/kernel_tests/video_test.py",
-    data = [
-        ":test_data",
-    ],
-    deps = [
-        ":video_ops_py",
-    ],
-    srcs_version = "PY2AND3",
-)
-
-py_library(
-    name = "video_py",
-    srcs = ([
-        "__init__.py",
-        "python/__init__.py",
-        "python/ops/__init__.py",
-    ]),
-    deps = [
-        ":video_ops_py"
-    ],
-    srcs_version = "PY2AND3",
 )
