+++ tests/test_hadoop.py
@@ -12,7 +12,7 @@
 # License for the specific language governing permissions and limitations under
 # the License.
 # ==============================================================================
-"""Tests for VideoDataset."""
+"""Tests for SequenceFileDataset."""
 
 from __future__ import absolute_import
 from __future__ import division
@@ -22,25 +22,30 @@ import os
 
 from tensorflow.python.platform import test
 
-from tensorflow_io.video.python.ops import video_dataset_ops
+from tensorflow_io.hadoop.python.ops import hadoop_dataset_ops
 from tensorflow.python.framework import constant_op
 from tensorflow.python.framework import dtypes
 from tensorflow.python.framework import errors
-from tensorflow.python.ops import image_ops
 from tensorflow.python.platform import resource_loader
 
 
-class VideoDatasetTest(test.TestCase):
+class SequenceFileDatasetTest(test.TestCase):
 
-  def test_video_dataset(self):
-    """Test case for VideoDataset."""
-    filename = os.path.join(resource_loader.get_data_files_path(),
-                            "testdata", "small.mp4")
+  def test_sequence_file_dataset(self):
+    """Test case for SequenceFileDataset.
+
+    The file is generated with `org.apache.hadoop.io.Text` for key/value.
+    There are 25 records in the file with the format of:
+    key = XXX
+    value = VALUEXXX
+    where XXX is replaced as the line number (starts with 001).
+    """
+    filename = os.path.join(os.path.dirname(os.path.abspath(__file__)), "test_hadoop", "string.seq")
 
     filenames = constant_op.constant([filename], dtypes.string)
     num_repeats = 2
 
-    dataset = video_dataset_ops.VideoDataset(filenames).repeat(
+    dataset = hadoop_dataset_ops.SequenceFileDataset(filenames).repeat(
         num_repeats)
     iterator = dataset.make_initializable_iterator()
     init_op = iterator.initializer
@@ -49,9 +54,10 @@ class VideoDatasetTest(test.TestCase):
     with self.cached_session() as sess:
       sess.run(init_op)
       for _ in range(num_repeats):  # Dataset is repeated.
-        for _ in range(166): # 166 frames
-          v = sess.run(get_next)
-          self.assertAllEqual(v.shape, (320, 560, 3))
+        for i in range(25):  # 25 records.
+          v0 = ("%03d" % (i + 1)).encode()
+          v1 = ("VALUE%03d" % (i + 1)).encode()
+          self.assertEqual((v0, v1), sess.run(get_next))
       with self.assertRaises(errors.OutOfRangeError):
         sess.run(get_next)
