+++ tests/test_hadoop/string.seq
