+++ tests/test_ignite/bin/start-plain.sh
