+++ tests/test_ignite/config/ignite-config-plain.xml
