+++ tests/test_ignite/stop_ignite.sh
