+++ tests/test_image/small-00.png
