+++ tests/test_image/small-01.png
