+++ tests/test_image/small-02.png
