+++ tests/test_image/small-03.png
