+++ tests/test_image/small-04.png
