+++ tests/test_image/small.tiff
