+++ tests/test_kafka/kafka_test.sh
