+++ tests/test_kinesis.py
