+++ tests/test_kinesis/kinesis_test.sh
